fix(warnings): Fix all 8 warnings from BUG_FIXES1_TODO (WARN-1 through WARN-8)
- WARN-1: Add bt_get_device_list_snapshot/bt_get_paired_devices_snapshot as
  safe copy-based alternatives to the raw-pointer accessors
- WARN-2: Log ESP_LOGW when bt_scan discovery buffer reaches 20-device limit
  instead of silently dropping devices
- WARN-3: Log ESP_LOGW with signal type when BtAppTask queue is full so queue
  backpressure is visible rather than looking like a hung task
- WARN-4: Initialize cmd_handle_i2s_config data[] buffer to '\0' so responses
  are never built from uninitialized memory
- WARN-5: Document serialization contract for mock state variables in
  cmd_handlers_bt.c
- WARN-6: Add prominent ESP_LOGW before NVS erase-on-version-mismatch so field
  teams can diagnose unexpected settings resets after firmware updates
- WARN-7: Change platform_storage_handle_t from uint32_t to uintptr_t so the
  host implementation can round-trip a pointer without 64-bit truncation
- WARN-8: Replace malloc/free in platform_sync_esp32.c with platform_malloc/
  platform_free to stay within the platform abstraction
- Bonus: fix pre-existing bt_scan.h #ifdef ESP_PLATFORM guard that hid
  bt_scan_handle_* declarations from host builds (causing implicit-declaration
  warnings); add forward declaration for bt_manager_test_set_connection_state
  in test_commands.c

55/55 host tests pass, 0 warnings.

+++ esp_bt_audio_source/components/platform_shim/platform_storage.h
@@ -40,11 +40,14 @@ extern "C" {
 
 /**
  * @brief Opaque handle for storage namespace
- * 
- * ESP32: nvs_handle_t
- * Host: Pointer to in-memory storage context
+ *
+ * ESP32: nvs_handle_t (uint32_t, fits in uintptr_t on 32-bit)
+ * Host:  Pointer to in-memory storage context cast to integer
+ *
+ * uintptr_t is used instead of uint32_t so the host implementation can round-
+ * trip a pointer through the handle without truncation on 64-bit systems.
  */
-typedef uint32_t platform_storage_handle_t;
+typedef uintptr_t platform_storage_handle_t;
 
 /**
  * @brief Storage access mode
